refactor: explorer refactoring
